Update image dater tool

+++ image_mover.py
@@ -1,11 +1,11 @@
-from typing import *
-import subprocess, re, os, datetime, argparse, sys, traceback, shutil
-q = "'"
-basefolder = 'C:/Backup'
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 
-# The following pattern matches with filenames that are already converted by the 'image_dater.py'
-# script.
-p = re.compile(r'(20\d\d)(\d\d)(\d\d)-(\d\d)(\d\d)(\d\d)-(\d+)')
+from typing import *
+import re, os, datetime, argparse, sys, shutil
+basefolder = '/mnt/c/Backup'
+
+
 
 def show_help() -> None:
     '''
@@ -17,16 +17,23 @@ def date_to_str(date_obj:datetime.datetime) -> str:
     '''
     Convert the date object to a string like:
     '20230601-184100'
+    This is the string format used for filenames by the 'image_dater.py' script (to which a counter
+    is appended).
     '''
     return date_obj.strftime('%Y%m%d-%H%M%S')
 
-def parse_images(dirpath):
+def move_images(dirpath:str, dry_run:bool=False) -> None:
     '''
-    Parse the given folder for images and move them.
+    Parse the given folder for images and move them if they're already converted by the
+    'image_dater.py' script.
     '''
+    # The following pattern matches with filenames that are already converted by the
+    # 'image_dater.py' script.
+    p = re.compile(r'(20\d\d)(\d\d)(\d\d)-(\d\d)(\d\d)(\d\d)-(\d+)')
+
     # First fill a list with all the files that are already converted by the 'image_dater.py'
     # script.
-    filepath_list = []
+    filepath_list:List[str] = []
     for root, dirs, files in os.walk(dirpath):
         for f in files:
             if not f.lower().endswith(('.heic', '.mov', '.jpeg', '.jpg', '.mp4', '.png')):
@@ -80,7 +87,7 @@ def parse_images(dirpath):
             while True:
                 dst_filepath = str(
                     f'{dst_folderpath}/'
-                    f'{date_to_str(filedate)}-{cntr:02d}.'
+                    f'{date_to_str(filedate)}-{cntr:03d}.'
                     f'{os.path.splitext(filename)}'
                 )
                 if os.path.exists(dst_filepath):
@@ -89,14 +96,16 @@ def parse_images(dirpath):
                 break
 
         # At this point, we know that the current file must be moved to 'dst_filepath' and the
-        # destination doesn't exist yet (the cntr has been taken care of). Move the file now.
+        # destination doesn't exist yet (the cntr ensured that). Move the file now.
         assert not os.path.exists(dst_filepath)
         if not os.path.isdir(dst_folderpath):
             print(f'CREATE: {dst_folderpath}')
-            os.makedirs(dst_folderpath)
+            if not dry_run:
+                os.makedirs(dst_folderpath)
         print(f'MOVE: {src_filepath} => {dst_filepath}')
         sys.stdout.flush()
-        shutil.move(src_filepath, dst_filepath)
+        if not dry_run:
+            shutil.move(src_filepath, dst_filepath)
         continue
     return
 
@@ -107,22 +116,38 @@ if __name__ == '__main__':
     )
     parser.add_argument('-h', '--help', action='store_true')
     parser.add_argument('-d', '--directory', action='store')
+    parser.add_argument('-n', '--dry-run',   action='store_true')
     args = parser.parse_args()
+
+    #& Sanitize input
+    if not args.help and not args.directory:
+        print('ERROR: No action specified')
+        show_help()
+        sys.exit(1)
+
+    #& Show help and quit
     if args.help:
         show_help()
+        print('\nQuit image mover tool\n')
+        sys.exit(0)
+
+    #& Parse directory and move images
     if args.directory is not None:
         if not os.path.isdir(args.directory):
-            print(f'ERROR: Cannot find directory: {q}{args.directory}{q}')
-            sys.exit(0)
+            print(f"ERROR: Cannot find directory: '{args.directory}'")
+            sys.exit(1)
         user_input = input(
-            f'Process the pictures from this folder:\n'
-            f'{q}{args.directory}{q}\n'
-            f'and store them at this location:\n'
-            f'{q}{basefolder}{q}\n'
-            f'Proceed? [yes|no]\n'
+            f"Process the pictures from this folder:\n"
+            f"'{args.directory}'\n"
+            f"and store them at this location:\n"
+            f"'{basefolder}'\n"
+            f"Proceed? [yes|no]\n"
         )
         if user_input.lower() in ('y', 'yes'):
-            parse_images(args.directory)
+            move_images(args.directory, args.dry_run)
+        print('\nQuit image mover tool\n')
+        sys.exit(0)
 
+    show_help()
     print('\nQuit image mover tool\n')
     sys.exit(0)
